requirements.txt
# Google AI/ADK Dependencies
google-adk>=0.1.0
google-genai>=0.3.0

# Google Cloud
google-cloud-aiplatform>=1.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0

# Development (optional)
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
